requirements.txt
pydantic
requests
beautifulsoup4
streamlit
mlflow

langchain-text-splitters
langchain-huggingface
langchain-community
langchain-chroma

chromadb
sentence-transformers

google-genai